avr: Program memory is organized as an array of bytes.
	- It was previously defined as an array of 16-bits words;
	- Several warnings have been fixed.

+++ avrsim/include/mcusim/avr/sim/simm8a.h
@@ -108,7 +108,7 @@ extern "C" {
 					   Rate Register */
 
 int MSIM_M8AInit(struct MSIM_AVR *mcu,
-		 uint16_t *pm, uint32_t pm_size,
+		 uint8_t *pm, uint32_t pm_size,
 		 uint8_t *dm, uint32_t dm_size);
 
 int MSIM_M8ALoadProgmem(struct MSIM_AVR *mcu, FILE *fp);
